fix(surprisal_backend): default-model dtype + resolved-dtype cache identity
Two reviewer P1 follow-ups on PR #93's first fix round:

1. Default-model fallback dropped dtype. _tier4_surprisal_block has
   two branches:
   (a) surprisal_model is not None: routes through
       _get_surprisal_backend, which now takes dtype. Covered by
       the first fix round.
   (b) surprisal_model is None: constructs
       SurprisalBackend(model_id=DEFAULT_MODEL) directly. The first
       fix round didn't touch this path, so an operator running
       --tier4 --surprisal-dtype fp32 *without* --surprisal-model
       still silently got "auto". Now passes
       dtype=surprisal_dtype to the direct construction.

2. Auto caches compared requested dtype, not resolved. Storing the
   request string lets an auto-on-CPU cache (resolved fp32)
   silently reuse on auto-on-H100 (resolved bf16). The cache compat
   check returned True because "auto" == "auto", even though the
   actual surprisal-series bit patterns differ.

   Fix: record both surprisal_dtype_requested (operator's request,
   possibly "auto") and surprisal_dtype_resolved (what
   _resolve_dtype picked on this host: "fp32" / "fp16" / "bf16",
   or None when torch isn't importable). cache_is_compatible
   compares the resolved labels.

   Field semantics in cache_is_compatible:
   - Field absent on cache: pre-1.93.0 cache, force re-score.
   - Field present and equal (including both None in no-torch
     environments): compatible.
   - Field present and unequal: re-score with a clear reason
     string showing the cached → current transition.
   - Tier-4 off: field meaningless, check skipped.

   The distinction between "field absent" and "field present and
   None" matters: a no-torch test environment writes None (torch
   not importable for resolution), and a later no-torch run of the
   same code should reuse the cache. A pre-1.93 cache has no field
   at all and must re-score.

Tests:

- test_tier4_default_model_fallback_passes_dtype pins the
  default-model branch passes dtype to the direct
  SurprisalBackend construction.
- test_cache_compat_invalidates_on_auto_crossing_hardware
  reproduces the original review case: both sides request "auto"
  but resolve to different actual precisions; cache must
  invalidate.
- test_cache_compat_auto_same_resolved_dtype_reuses: positive
  mirror — same auto resolution on both sides reuses cleanly.
- test_cache_compat_both_resolved_none_no_torch_is_compatible
  pins the no-torch test environment's both-None case as
  compatible (distinct from the field-absent case).
- Existing tests updated to use the new field names
  (surprisal_dtype_requested / _resolved). _resolve_surprisal_dtype_label
  is mocked in tests so they don't depend on host torch state.

85 tests pass across the affected suites (was 81, +4 new
including the both-None case). 2 pre-existing embedding failures
unrelated to dtype work remain.

surprisal_audit.py standalone path is still per-text non-batched
and non-dtype-aware; chunk 2 covers that.

+++ plugins/setec-voiceprint/scripts/variance_audit.py
@@ -874,11 +874,17 @@ def _tier4_surprisal_block(
         else:
             try:
                 # Construct a default backend lazily. Same lazy-load
-                # semantics as the audit script's CLI path.
+                # semantics as the audit script's CLI path. Pass dtype
+                # explicitly — without this the per-entry fallback with
+                # no --surprisal-model would silently load the default
+                # at the SurprisalBackend "auto" default regardless of
+                # operator intent (reviewer P1 follow-up on PR #93).
                 from surprisal_backend import (  # type: ignore
                     DEFAULT_MODEL, SurprisalBackend,
                 )
-                backend = SurprisalBackend(model_id=DEFAULT_MODEL)
+                backend = SurprisalBackend(
+                    model_id=DEFAULT_MODEL, dtype=surprisal_dtype,
+                )
             except ImportError as exc:
                 return {
                     "available": False,
